example_plugins/grayscale_python/requirements.txt
grpcio
numpy
